requirements.txt
sqlalchemy>=2.0
psycopg2-binary>=2.9
pandas>=2.0
numpy>=1.24
streamlit>=1.30
streamlit-folium>=0.15
folium>=0.15
requests>=2.31
tenacity>=8.2
python-dotenv>=1.0
click>=8.0
tqdm>=4.65
pytest>=7.0
